requirements.txt
Flask~=3.1.2
pip~=23.2.1
Jinja2~=3.1.6
typing_extensions~=4.15.0
SQLAlchemy~=2.0.46
openpyxl~=3.1.5
Flask-SQLAlchemy~=3.1.1
python-docx ~= 1.2.0